changed the handling of the logout to add it to the dashboard for the centralization of the socket
