+++ chathaven/src/components/LogoutButton.js
@@ -1,22 +1,7 @@
-import { useRouter } from "next/router";
 import "./styles/Dashboard.css";
-import { useSocket } from "./SocketContext";
 import { CiLogout } from "react-icons/ci";
 
-export default function LogoutButton() {
-  const router = useRouter();
-  const { socket, userId, updateStatus } = useSocket();
-
-  const handleLogout = async () => {
-    // Emit status change to "unavailable" before logging out
-    if (socket && userId) {
-      console.log("Setting status to 'unavailable' for user:", userId);
-      updateStatus("unavailable"); // Set the status to 'unavailable'
-    }
-    await fetch("/api/auth/logout", { method: "POST", credentials: "include" });
-    router.push("/login");
-  };
-
+export default function LogoutButton({ handleLogout }) {
   return (
     <button
       id="logoutButton"
